5_ui/requirements.txt
altair==4.2.0
attrs==22.2.0
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
filelock==3.9.0
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.2
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
numpy==1.24.1
nltk==3.8.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.14.0
pymongo==4.3.3
Pympler==1.0.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.2.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
snscrape==0.5.0.20230113
soupsieve==2.3.2.post1
streamlit==1.17.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
zipp==3.11.0
langdetect==1.0.9

sklearn-pandas==2.2.0
pickleshare==0.7.5
cloudpickle==2.2.1
jsonpickle==3.0.1
xgboost==1.7.5

plotly==5.14.1
elasticsearch==8.6.2